requirements.txt
# Core ML/DL
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Language detection
langdetect>=1.0.9

# LLM APIs
openai>=1.0.0
anthropic>=0.25.0

# Video processing (for main.py end-to-end pipeline)
opencv-python>=4.8.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Evaluation & plotting
matplotlib>=3.7.0
seaborn>=0.12.0
